// Qanat.Web/src/app/shared/generated/model/record-set-schema.ts
export class RecordSetSchema { 
    FieldSchemata: Array<FieldSchema>;
    GroupSchemata?: Array<GroupSchema> | null;
    RecordNotMeasurable?: NotMeasurableSchema;
    AdditionalProperties?: { [key: string]: object | null; } | null;
    Name: string;
    CanonicalName: string;
    Description?: string | null;
    MinCount: number;
    MaxCount?: number | null;
    RecordSetNotMeasurable?: NotMeasurableSchema;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}